Version File to check against for updates
Version file for servers running lime cart  to check against for version
changes, if change is detected suggest a upgrade to user in admin panel
or in email. Use Git to host version file, access with curl and raw git
data ?? Must be updated for major version releases and security patches,
minor changes not so important.